requirements.txt
numpy
torch
pillow
matplotlib
sklearn
tqdm
python-fcl
seaborn
pyyaml